Popravki algoritma MaxSum

+++ AlgorithmsRST/Concepts/DivideAndConquer.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -53,12 +54,12 @@ namespace Borut.Lectures.AlgorithmsRST
         /// Given a list of integers, it finds the maximum sum of contiguous elements
         /// in O(n^2) time.
         /// </summary>
-        public static int NaiveMaxSubsequenceSum(List<int> lst)
+        public static long NaiveMaxSubsequenceSum(List<int> lst)
         {
-            int max = int.MinValue;
+            long max = long.MinValue;
             for (int i = 0; i < lst.Count; i++)
             {
-                int sumI = 0;
+                long sumI = 0;
                 for (int j = i; j < lst.Count; j++)
                 {
                     sumI += lst[j];
@@ -74,7 +75,7 @@ namespace Borut.Lectures.AlgorithmsRST
         /// Given a list of integers, it finds the maximum sum of contiguous elements
         /// in O(n log(n)) time using divide and conquer approach.
         /// </summary>
-        public static int MaxSubsequenceSum_DivAndCon(List<int> lst, int start, int end)
+        public static long MaxSubsequenceSum_DivAndCon(List<int> lst, int start, int end)
         {
             // Check if we only have one element
             if (start == end)
@@ -83,7 +84,7 @@ namespace Borut.Lectures.AlgorithmsRST
             int half = (start + end) / 2;
 
             // Check left side from the break
-            int leftMax = int.MinValue, leftSum = 0;
+            long leftMax = int.MinValue, leftSum = 0;
             for (int i = half; i >= start; i--)
             {
                 leftSum += lst[i];
@@ -94,14 +95,14 @@ namespace Borut.Lectures.AlgorithmsRST
             }
 
             // Check right side from the break
-            int rightMax = int.MinValue, rightSum = 0;
+            long rightMax = int.MinValue, rightSum = 0;
             for (int i = half + 1; i <= end; i++)
             {
                 rightSum += lst[i];
                 if (rightMax < rightSum) rightMax = rightSum;
             }
 
-            return new List<int>()
+            return new List<long>()
                 {
                     leftMax + rightMax,
                     MaxSubsequenceSum_DivAndCon(lst, start, half),
@@ -131,10 +132,10 @@ namespace Borut.Lectures.AlgorithmsRST
         /// Given a list of integers, it finds the maximum sum of contiguous elements
         /// in linear time.
         /// </summary>
-        public static int MaxSubsequenceSumLinearWorkingForAllNegative(List<int> lst)
+        public static long MaxSubsequenceSumLinearWorkingForAllNegative(List<int> lst)
         {
-            int max = int.MinValue;
-            int currentMax = 0;
+            long max = int.MinValue;
+            long currentMax = 0;
             bool allNegativeValues = true;
             for (int i = 0; i < lst.Count; i++)
             {
